Merge Test to Stable

+++ src/metax_api/models/catalog_record.py
@@ -44,7 +44,7 @@ ACCESS_TYPES = {
 LEGACY_CATALOGS = settings.LEGACY_CATALOGS
 IDA_CATALOG = settings.IDA_DATA_CATALOG_IDENTIFIER
 PAS_CATALOG = settings.PAS_DATA_CATALOG_IDENTIFIER
-
+DFT_CATALOG = settings.DFT_DATA_CATALOG_IDENTIFIER
 
 class DiscardRecord(Exception):
     pass
@@ -956,36 +956,37 @@ class CatalogRecord(Common):
             super(Common, self).delete()
             return
 
-        if self.has_alternate_records():
-            self._remove_from_alternate_record_set()
-        if get_identifier_type(self.preferred_identifier) == IdentifierType.DOI:
-            self.add_post_request_callable(DataciteDOIUpdate(self, self.research_dataset['preferred_identifier'],
-                                                             'delete'))
+        elif self.state == self.STATE_PUBLISHED:
+            if self.has_alternate_records():
+                self._remove_from_alternate_record_set()
+            if get_identifier_type(self.preferred_identifier) == IdentifierType.DOI:
+                self.add_post_request_callable(DataciteDOIUpdate(self, self.research_dataset['preferred_identifier'],
+                                                                'delete'))
 
-        if self._dataset_has_rems_managed_access() and settings.REMS['ENABLED']:
-            self._pre_rems_deletion('dataset deletion')
-            super().save(update_fields=['rems_identifier', 'access_granter'])
+            if self._dataset_has_rems_managed_access() and settings.REMS['ENABLED']:
+                self._pre_rems_deletion('dataset deletion')
+                super().save(update_fields=['rems_identifier', 'access_granter'])
 
-        self.add_post_request_callable(RabbitMQPublishRecord(self, 'delete'))
+            self.add_post_request_callable(RabbitMQPublishRecord(self, 'delete'))
 
-        log_args = {
-            'event': 'dataset_deleted',
-            'user_id': self.user_modified,
-            'catalogrecord': {
-                'identifier': self.identifier,
-                'preferred_identifier': self.preferred_identifier,
-                'data_catalog': self.data_catalog.catalog_json['identifier'],
+            log_args = {
+                'event': 'dataset_deleted',
+                'user_id': self.user_modified,
+                'catalogrecord': {
+                    'identifier': self.identifier,
+                    'preferred_identifier': self.preferred_identifier,
+                    'data_catalog': self.data_catalog.catalog_json['identifier'],
+                }
             }
-        }
-        if self.catalog_is_legacy():
-            # delete permanently instead of only marking as 'removed'
-            super().delete()
-        else:
-            super().remove(*args, **kwargs)
-            log_args['catalogrecord']['date_removed'] = datetime_to_str(self.date_removed)
-            log_args['catalogrecord']['date_modified'] = datetime_to_str(self.date_modified)
+            if self.catalog_is_legacy():
+                # delete permanently instead of only marking as 'removed'
+                super().delete()
+            else:
+                super().remove(*args, **kwargs)
+                log_args['catalogrecord']['date_removed'] = datetime_to_str(self.date_removed)
+                log_args['catalogrecord']['date_modified'] = datetime_to_str(self.date_modified)
 
-        self.add_post_request_callable(DelayedLog(**log_args))
+            self.add_post_request_callable(DelayedLog(**log_args))
 
     def deprecate(self, timestamp=None):
         self.deprecated = True
@@ -1052,6 +1053,12 @@ class CatalogRecord(Common):
 
     def catalog_is_pas(self):
         return self.data_catalog.catalog_json['identifier'] == PAS_CATALOG
+
+    def catalog_is_dft(self):
+        return self.data_catalog.catalog_json['identifier'] == DFT_CATALOG
+
+    def is_published(self):
+        return self.state == self.STATE_PUBLISHED
 
     def has_alternate_records(self):
         return bool(self.alternate_record_set)
@@ -1247,7 +1254,8 @@ class CatalogRecord(Common):
 
             super().save()
 
-            self.add_post_request_callable(RabbitMQPublishRecord(self, 'create'))
+            if self.state == self.STATE_PUBLISHED:
+                self.add_post_request_callable(RabbitMQPublishRecord(self, 'create'))
 
         _logger.info(
             'Created a new <CatalogRecord id: %d, '
@@ -1439,7 +1447,8 @@ class CatalogRecord(Common):
             self.add_post_request_callable(DataciteDOIUpdate(self, self.research_dataset['preferred_identifier'],
                                                              'update'))
 
-        self.add_post_request_callable(RabbitMQPublishRecord(self, 'update'))
+        if self.state == self.STATE_PUBLISHED:
+            self.add_post_request_callable(RabbitMQPublishRecord(self, 'update'))
 
         log_args = {
             'event': 'dataset_updated',
@@ -2016,8 +2025,8 @@ class CatalogRecord(Common):
             self.add_post_request_callable(DataciteDOIUpdate(new_version,
                                                              new_version.research_dataset['preferred_identifier'],
                                                              'create'))
-
-        new_version.add_post_request_callable(RabbitMQPublishRecord(new_version, 'create'))
+        if self.state == self.STATE_PUBLISHED:
+            new_version.add_post_request_callable(RabbitMQPublishRecord(new_version, 'create'))
 
         old_version.new_dataset_version_created = new_version.identifiers_dict
         old_version.new_dataset_version_created['version_type'] = 'dataset'
@@ -2451,7 +2460,8 @@ class CatalogRecord(Common):
 
             super().save()
 
-        self.add_post_request_callable(RabbitMQPublishRecord(self, 'update'))
+        if self.state == self.STATE_PUBLISHED:
+            self.add_post_request_callable(RabbitMQPublishRecord(self, 'update'))
 
         return True if new_state == self.CUMULATIVE_STATE_YES else False
 
@@ -2506,7 +2516,8 @@ class CatalogRecord(Common):
             self._create_new_dataset_version()
 
         super().save()
-        self.add_post_request_callable(RabbitMQPublishRecord(self, 'update'))
+        if self.state == self.STATE_PUBLISHED:
+            self.add_post_request_callable(RabbitMQPublishRecord(self, 'update'))
 
         return (self.cumulative_state != self.CUMULATIVE_STATE_YES, len(added_file_ids))
